Fix for safari issues, seems a lot slower than before though.
