+++ index.html
@@ -114,6 +114,7 @@
 <script src="examples/imageLoader.js"></script>
 
 <script>
+
 	const cTools = cornerstoneTools.init();
 	const imageIds = [
 		'example://1',
