+++ src/tools/BrushTool.js
@@ -141,50 +141,3 @@ export default class BrushTool extends BaseBrushTool {
     return (eventData.event && eventData.event.ctrlKey) || eventData.ctrlKey;
   }
 }
-
-/*
-function defaultBrushToolConfiguration () {
-  return {
-    keyBinds: {
-      increaseBrushSize: '+',
-      decreaseBrushSize: '-',
-      nextSegmentation: ']',
-      previousSegmentation: '['
-    }
-  };
-}
-*/
-
-/* Safari and Edge polyfill for createImageBitmap
- * https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/createImageBitmap
- */
-
-if (!('createImageBitmap' in window)) {
-  window.createImageBitmap = async function (imageData) {
-    return new Promise((resolve) => {
-      const img = document.createElement('img');
-
-      img.addEventListener('load', function () {
-        resolve(this);
-      });
-
-      const conversionCanvas = document.createElement('canvas');
-
-      conversionCanvas.width = imageData.width;
-      conversionCanvas.height = imageData.height;
-
-      const conversionCanvasContext = conversionCanvas.getContext('2d');
-
-      conversionCanvasContext.putImageData(
-        imageData,
-        0,
-        0,
-        0,
-        0,
-        conversionCanvas.width,
-        conversionCanvas.height
-      );
-      img.src = conversionCanvas.toDataURL();
-    });
-  };
-}
